Fix misdeclaration of DECLARE_BACKEND unlockAchievement()
A-ha! Previously, STEAM_unlockAchievement() (for example) was being
forward-declared as an argument that took in any arguments and discarded
them all, which compiled fine because the actual function took in an
argument, and we called it with an argument. But when I changed it to
explicitly declare the void, it resulted in a compile error, revealing
the fact that there was a mismatch between the forward-declaration and
the actual implementation.

This is fixed now.